+++ packages/vue/src/components/collection/collection.stories.vue
@@ -1,0 +1,11 @@
+<script setup>
+import ListSelection from './examples/list-selection.vue'
+</script>
+
+<template>
+  <Story title="Collection">
+    <Variant title="ListSelection">
+      <ListSelection />
+    </Variant>
+  </Story>
+</template>